Adjusted TestingSqliteConnectionFactory to provide sharable in-memory connections.
To carry out Unit Tests which rely on SQLite, we use an in-memory database to perform operations on, which is deleted when the connection made to it is closed (https://www.sqlite.org/inmemorydb.html
https://learn.microsoft.com/en-us/dotnet/standard/data/sqlite/in-memory-databases). This is fine for testing one-off operations, as they are not dependent on the state of the database or its existing contents, but would become problematic once
other operations need to be ran as part of testing setup, as a method which closes a connection would inadvertantly wipe the in-memory database. To prevent this, TestingSqliteConnectionFactory has been adjusted to maintain a connection that is
shared with other test subjects that use it throughout its lifetime, which should prevent this.