Added Interface and Stub class for initializing sessions.
To allow a client built on top of AcaiCore to begin working from a journal file, the goal is to use a Session Initialization Facade to create a Session (which will encapsulate operations to other components such as Gateways) which the client can then use from either an existing journal file or by creating a new one. In service to this, an Interface for the Facade and stub class for the Session have been created respectively.