Created updated FoodJournalPage which closer follows MVVM (and added Maui Community Toolkit MVVM as Dependency)
A new Page has been created (FoodJournalPage) which will eventually replace the old MainPage. This has been rewritten with assistance from beginner materials in order to closer bring it to a MVVM pattern. If I've understood it correctly, the .XAML can act as a view. As for the View-Model side, a dedicated class has been created for this, which is now provided with the "View" as a Singleton using Dependency Injection. As for the "Model", this should arguably be fulfilled by the use of AcaiCore to encapsulate critical business logic which is covered by Unit Tests. To assist with this entire process, the MAUI Community Toolkit Package (MVVM) has been added as a dependency.