Removing iOS and Mac Catalyst as AcaiMobile Target and explicit versioning of Microsoft.Maui.Controls
As part of dealing with growing pains to get AcaiMobile building successfully from an Ubuntu machine, I've taken out iOS and Mac Catalyst as Targets. This was something already long overdue as there are not plans to target this platform anyway and reducing the friction of setting up mobile workflows is always a plus.

Additionally, to resolve version downgrade errors and changes in .NET 8 that would otherwise block the build, Microsoft.Maui.Controls has been explicitly included as a PackageReference as <UseMaui> no longer includes those packages automatically in the project.