Refactor: Adjusted PresentInConnection method in FoodItemTableSchema to use PRAGMA queries
Because we want to extend the food_items table to include new columns using a new IJournalTableSchema, the current approach of checking for the exact creation query used in sqlite_master no longer feels sustainable without having to retroactively change old schemas when new ones are introduced, as altering a table will also alter the creation query returned.

In response to this, PRAGMA (TABLE_INFO) queries will now be used to validate the presence of all required columns in the schema, the idea being that more granular control can be given to ensuring that only the relevant columns introduced by a schema can be validated.